Don't use a GtkDialog without a transient parent
In our human-evaluation tests, we presented the rendered RTF in a
GtkDialog without a transient parent. This is discouraged in GTK 3, so
instead use a GtkWindow and run a main loop instead of gtk_dialog_run().
Instead of responses, we use signal handlers and pass in a boolean
variable to the handler for clicks on the Yes button.

This also gets rid of stock buttons, which are deprecated in GTK 3.