Import RTF from file in tests
Some of the testcases contain references to external images, so we should
not import them from a string; in that case the references are only
correct if the current working directory is the same as the path where
the testcase resides. Instead, import the testcases with GFile so the
external references can be resolved correctly.

Add a note to the documentation explaining this.