Handle Clang in the ignored warning in plist.[ch]
We ignored a GCC warning about "const" having no effect because
including "const" in the declaration does have an effect on the
introspection scanner.

Clean up the preprocessor code a bit and make it emit the #pragmas for
Clang as well, which understands the same syntax as GCC. And make
gtk-doc ignore it.